Clarify provisioning errors, especially regarding unsupported token models
Per #67, an incorrect token-model can result in the provisioning response
returning (result='Malformed request', reasoncode='4E00'). It appears that
this particular response only occurs if the token model is NOT an
alphanumeric string of 3-4 characters. We can specifically check this
client-side.

If the token model requested IS an alphanumeric string of 3-4 characters,
but not a supported one, then we get the more comprehensible
(result='Unsupported token model', reasoncode='4E0D').

Also, list the old/new pairs of token models a little more clearly in the
README.